[Fix] 회원가입 기능 : 중복되는 예외처리 제거, 불필요한 DB 쿼리 수정
GlobalExceptionHandler를 못 보고 예외처리를 따로 했던 부분을 제거하고 이메일, 비밀번호 유효성 검사를 먼저 하고 그 뒤에 중복 이메일을 검사하여 불필요한 쿼리가 발생하지 않게 수정했습니다.
