+++ src/main/java/com/api/domain/auth/service/AuthService.java
@@ -1,14 +1,13 @@
-package com.api.domain.users.service;
+package com.api.domain.auth.service;
 
 import com.api.domain.config.JwtUtil;
 import com.api.domain.users.dto.UserRequestDto;
 import com.api.domain.users.dto.UserResponseDto;
 import com.api.domain.users.entity.User;
 import com.api.domain.users.repository.UserRepository;
+import com.api.domain.users.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -29,13 +28,13 @@ public class AuthService {
         if(!userService.checkEmailPattern(requestDto.getEmail())) {
             throw new IllegalArgumentException("이메일 형식이 아닙니다.");
         }
-        // 중복된 이메일 검증
-        if (userRepository.findByEmail(requestDto.getEmail()).isPresent()) {
-            throw new IllegalArgumentException("중복된 이메일입니다.");
-        }
         // 비밀번호 형식 검증
         if(!userService.PasswordPattern(requestDto.getPassword())) {
             throw new IllegalArgumentException("영어 대소문자, 숫자, 특수문자가 모두 포함되어야합니다.");
+        }
+        // 중복된 이메일 검증
+        if (userRepository.findByEmail(requestDto.getEmail()).isPresent()) {
+            throw new IllegalArgumentException("중복된 이메일입니다.");
         }
 
         //유저 저장
